1-update-spectrum-index-pattern - FEAT - Replace general pattern for spectrum id index to allow any characters id_file_parser.py - Remove integer conversion of spectrum id index in percolator.py

+++ ms2rescore/percolator.py
@@ -265,7 +265,8 @@ class PercolatorIn:
         """Get spectrum index column, parsed from Percolator SpecId."""
         if not pattern:
             pattern = r".+_([0-9]+)_[0-9]+_[0-9]+"
-        id_col = self.df["SpecId"].str.extract(pattern, expand=False).astype(int)
+        #id_col = self.df["SpecId"].str.extract(pattern, expand=False).astype(int)
+        id_col = self.df["SpecId"].str.extract(pattern, expand=False)
         if id_col.duplicated().any():
             raise PercolatorInError("Issue in matching spectrum IDs, duplicates found.")
         if id_col.isna().any():
